rafstore: address comments
